requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
firebase-admin>=6.4.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0
httpx>=0.27.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.29.0
pillow>=10.0.0
plotly>=5.18.0
pandas>=2.0.0
